Update GoogleCloudVision Key Regex
If there is an underscore in your API key for Google Cloud Vision. The Regex only matches everything BEFORE the underscore, and since Google nowadays also allows for underscore in the API key it should be added to the regex as well. 

Very small change to make the whole script work again.